Finished animation
I followed simple tutorials to learn everything I needed to do here. This includes lighting, bell destruction and using Unreal's Sequencer.

Thanks to my prior knowledge of materials and the basic assets given by Unreal, it was easy putting the scene together and adapt the materials to the render in Blender.